requirements.txt
streamlit==1.35.0
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.6
langchain-openai==0.1.14
faiss-cpu==1.8.0
sentence-transformers==3.0.1
PyMuPDF==1.24.5
pandas==2.2.2
beautifulsoup4==4.12.3
requests==2.32.3
deep-translator==1.11.4
python-dotenv==1.0.1
langdetect==1.0.9
nltk==3.8.1
plotly==5.22.0
openai==1.35.0
groq==0.9.0
